requirements.txt

PyPDF2==3.0.1
SQLAlchemy==2.0.25
SpeechRecognition
beautifulsoup4==4.12.3
bs4==0.0.2
cachetools==5.3.2
celery
colorlog==6.8.2
dataclasses-json==0.6.3
duckduckgo_search==4.2
fake-useragent==1.5.1
fsspec
fuzzywuzzy==0.18.0
gTTS
googlesearch-python
gunicorn

jsonpatch==1.33
jsonpath-python==1.0.6


langchain-community==0.2.3
langchain-core

langchain-openai==0.1.8
langchain-pinecone
langchain-text-splitters
langchain==0.2.8 
langchain_community
langchainhub
langgraph==0.0.64
langsmith
numpy==1.26.3


openai
pandas

pdf2image==1.17.0
pikepdf
pinecone-client
playwright==1.45.0
praw==7.7.1
psutil
pyTelegramBotAPI
pydantic==2.5.3
pydantic_core==2.14.6
pydub

pypdf==3.17.4
python-dotenv

redis
sentence-transformers
starlette==0.35.1

tavily-python 
termcolor==2.4.0
threadpoolctl==3.2.0

tiktoken
timm==0.9.12
tokenizers==0.15.0
unstructured-client==0.15.2
unstructured-inference==0.7.21
unstructured.pytesseract==0.3.12
unstructured==0.12.0

supervisor
flask